backend/requirements.txt
Django==3.2.15
djangorestframework==3.13.1
djoser==2.1.0
webcolors==1.12
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
Pillow==9.2.0
